// features/profiles/schemas/projectSchema.ts
import * as z from "zod";

export const projectSchema = z.object({
  title: z.string().min(3, "Title must be at least 3 characters"),
  description: z.string().min(10, "Description must be at least 10 characters"),
  liveLink: z.string().url("Must be a valid URL").optional().or(z.literal("")),
  repoLink: z.string().url("Must be a valid URL").optional().or(z.literal("")),
});

export type ProjectFormValues = z.infer<typeof projectSchema>;